bigquery_scripts/requirements.txt
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
